feat: LottoApplication 당첨 통계 수동 구매 티켓 포함

+++ src/main/kotlin/lotto/domain/Money.kt
@@ -1,7 +1,7 @@
 package lotto.domain
 
 @JvmInline
-value class Money(val value: Int) {
+value class Money(val value: Long) {
     init {
         require(value >= 0) { "$value 금액은 설정할수 없습니다. 0 이상에 금액을 입력해주세요" }
     }
