+++ src/main/kotlin/lotto/domain/TicketSeller.kt
@@ -5,7 +5,7 @@ import lotto.util.RandomGenerate
 class TicketSeller(private val ticketPrice: Money, private val randomGenerate: RandomGenerate) {
 
     fun buyPossibleLottoTicket(userMoney: Money): List<LottoTicket> {
-        return List(userMoney.value / ticketPrice.value) {
+        return List((userMoney.value / ticketPrice.value).toInt()) {
             LottoTicket.ofInts(
                 randomGenerate.makeRandomUniqueIntList(
                     LottoTicketNumbers.LOTTO_TICKET_NUMBER_SIZE,
